chore: issue 패키지 이름 변경
- issue -> board
- 예외 클래스 분리
