+++ api/src/main/java/com/opensourcereader/api/controller/board/request/IssueGetRequest.java
@@ -1,0 +1,3 @@
+package com.opensourcereader.api.controller.board.request;
+
+public record IssueGetRequest(Long repositoryId) {}
