--- api/src/main/java/com/opensourcereader/api/controller/issue/request/IssueGetRequest.java
@@ -1,3 +1,0 @@
-package com.opensourcereader.api.controller.issue.request;
-
-public record IssueGetRequest(Long repositoryId) {}
