+++ core/src/main/java/com/opensourcereader/core/board/dto/IssueCommentDto.java
@@ -1,0 +1,3 @@
+package com.opensourcereader.core.board.dto;
+
+public record IssueCommentDto(String body) {}
