+++ core/src/main/java/com/opensourcereader/core/board/dto/IssueDto.java
@@ -1,0 +1,3 @@
+package com.opensourcereader.core.board.dto;
+
+public record IssueDto(Long tagId, Boolean isOpened, String title, String body) {}
