+++ core/src/main/java/com/opensourcereader/core/board/dto/LabelDto.java
@@ -1,0 +1,3 @@
+package com.opensourcereader.core.board.dto;
+
+public record LabelDto(String name, String description, String color) {}
