+++ core/src/main/java/com/opensourcereader/core/board/entity/IssueComment.java
@@ -1,0 +1,40 @@
+package com.opensourcereader.core.board.entity;
+
+import com.opensourcereader.core.BaseEntity;
+import com.opensourcereader.core.user.entity.User;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "IssueComments")
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+public class IssueComment extends BaseEntity {
+
+  @Column(nullable = false)
+  private Long githubId;
+
+  @OneToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "author_id", nullable = false)
+  private User author;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "issue_id", nullable = false)
+  private Issue issue;
+
+  @Column(nullable = false)
+  private String body;
+}
