+++ core/src/main/java/com/opensourcereader/core/board/exception/IssueNotFoundException.java
@@ -1,0 +1,8 @@
+package com.opensourcereader.core.board.exception;
+
+public class IssueNotFoundException extends BoardException {
+
+  public IssueNotFoundException() {
+    super(BoardErrorCode.ISSUE_NOT_FOUND);
+  }
+}
