+++ core/src/main/java/com/opensourcereader/core/board/repository/LabelRepository.java
@@ -1,0 +1,15 @@
+package com.opensourcereader.core.board.repository;
+
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.opensourcereader.core.board.entity.Issue;
+import com.opensourcereader.core.board.entity.Label;
+
+public interface LabelRepository extends JpaRepository<Label, Long> {
+
+  List<Label> findAllByRepositoryId(Long repositoryId);
+
+  List<Label> findAllByIssue(Issue issue);
+}
