+++ core/src/main/java/com/opensourcereader/core/board/service/IssueCommentService.java
@@ -1,0 +1,10 @@
+package com.opensourcereader.core.board.service;
+
+import java.util.List;
+
+import com.opensourcereader.core.board.entity.IssueComment;
+
+public interface IssueCommentService {
+
+  List<IssueComment> findAllByIssueId(Long issueId);
+}
