--- core/src/main/java/com/opensourcereader/core/issue/dto/IssueCommentDto.java
@@ -1,3 +1,0 @@
-package com.opensourcereader.core.issue.dto;
-
-public record IssueCommentDto(String body) {}
