--- core/src/main/java/com/opensourcereader/core/issue/dto/IssueDto.java
@@ -1,3 +1,0 @@
-package com.opensourcereader.core.issue.dto;
-
-public record IssueDto(Long tagId, Boolean isOpened, String title, String body) {}
