--- core/src/main/java/com/opensourcereader/core/issue/dto/LabelDto.java
@@ -1,3 +1,0 @@
-package com.opensourcereader.core.issue.dto;
-
-public record LabelDto(String name, String description, String color) {}
