--- core/src/main/java/com/opensourcereader/core/issue/exception/IssueException.java
@@ -1,15 +1,0 @@
-package com.opensourcereader.core.issue.exception;
-
-import com.opensourcereader.core.exception.OpenSourceReaderException;
-
-import lombok.Getter;
-
-@Getter
-public abstract class IssueException extends OpenSourceReaderException {
-
-  private final IssueErrorCode issueErrorCode;
-
-  public IssueException(IssueErrorCode issueErrorCode) {
-    this.issueErrorCode = issueErrorCode;
-  }
-}
