--- core/src/main/java/com/opensourcereader/core/issue/exception/IssueNotFoundException.java
@@ -1,8 +1,0 @@
-package com.opensourcereader.core.issue.exception;
-
-public class IssueNotFoundException extends IssueException {
-
-  public IssueNotFoundException() {
-    super(IssueErrorCode.ISSUE_NOT_FOUND);
-  }
-}
