--- core/src/main/java/com/opensourcereader/core/issue/repository/IssueCommentRepository.java
@@ -1,12 +1,0 @@
-package com.opensourcereader.core.issue.repository;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.opensourcereader.core.issue.entity.IssueComment;
-
-public interface IssueCommentRepository extends JpaRepository<IssueComment, Long> {
-
-  List<IssueComment> findAllByIssueId(Long issueId);
-}
