--- core/src/main/java/com/opensourcereader/core/issue/repository/LabelRepository.java
@@ -1,15 +1,0 @@
-package com.opensourcereader.core.issue.repository;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.opensourcereader.core.issue.entity.Issue;
-import com.opensourcereader.core.issue.entity.Label;
-
-public interface LabelRepository extends JpaRepository<Label, Long> {
-
-  List<Label> findAllByRepositoryId(Long repositoryId);
-
-  List<Label> findAllByIssue(Issue issue);
-}
