--- core/src/main/java/com/opensourcereader/core/issue/service/IssueCommentService.java
@@ -1,10 +1,0 @@
-package com.opensourcereader.core.issue.service;
-
-import java.util.List;
-
-import com.opensourcereader.core.issue.entity.IssueComment;
-
-public interface IssueCommentService {
-
-  List<IssueComment> findAllByIssueId(Long issueId);
-}
