--- core/src/main/java/com/opensourcereader/core/issue/service/IssueCommentServiceImpl.java
@@ -1,22 +1,0 @@
-package com.opensourcereader.core.issue.service;
-
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-
-import com.opensourcereader.core.issue.entity.IssueComment;
-import com.opensourcereader.core.issue.repository.IssueCommentRepository;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class IssueCommentServiceImpl implements IssueCommentService {
-
-  private final IssueCommentRepository issueCommentRepository;
-
-  @Override
-  public List<IssueComment> findAllByIssueId(Long issueId) {
-    return issueCommentRepository.findAllByIssueId(issueId);
-  }
-}
